Add multiply transition stages to manage regular vault, extract icons to separate file
