+++ theme/icons.tsx
@@ -1,0 +1,971 @@
+import { icons as daiUiIcons } from '@makerdao/dai-ui-icons'
+// @ts-ignore
+import { icons as brandingIcons } from '@makerdao/dai-ui-icons-branding'
+import React from 'react'
+
+import { tokenIcons } from './tokenIcons'
+
+const customIcons = {
+  chevron: {
+    path: (
+      <path
+        d="M9.61862 14.5513C9.81829 14.7864 10.1811 14.7864 10.3808 14.5513L16.0944 7.82366C16.3703 7.49888 16.1394 7 15.7133 7H4.28612C3.86001 7 3.62918 7.49888 3.90501 7.82366L9.61862 14.5513Z"
+        fill="black"
+      />
+    ),
+    viewBox: '0 0 20 20',
+  },
+  arrow_down: {
+    path: (
+      <path
+        d="M8.5498 0.727272H6.61088V12.4503L2.129 7.96839L0.80158 9.31072L7.58034 16.0895L14.374 9.31072L13.0168 7.96839L8.5498 12.4503V0.727272Z"
+        fill="currentColor"
+      />
+    ),
+    viewBox: '0 0 15 17',
+  },
+  arrow_up: {
+    path: (
+      <path
+        d="M10.1007 22.7273H12.5346L12.5346 10.4677L17.182 15.1151L18.8482 13.4325L11.3176 5.90199L3.77077 13.4325L5.46964 15.1151L10.1007 10.4677L10.1007 22.7273Z"
+        fill="currentColor"
+      />
+    ),
+    viewBox: '0 0 28 28',
+  },
+  arrow_up_thin: {
+    path: (
+      <path
+        d="M6.4502 16.2727H8.38912L8.38912 4.54972L12.871 9.03161L14.1984 7.68928L7.41966 0.910511L0.625977 7.68928L1.98322 9.03161L6.4502 4.54972L6.4502 16.2727Z"
+        fill="currentColor"
+      />
+    ),
+    viewBox: '0 0 15 17',
+  },
+  arrow_left: {
+    path: (
+      <path
+        d="M12.1587 22.6768L13.7773 21.0581L5.26514 12.5649H28.9165V10.2988H5.26514L13.7773 1.80566L12.1587 0.206055L0.961426 11.4414L12.1587 22.6768Z"
+        fill="currentColor"
+      />
+    ),
+    viewBox: '0 0 29 23',
+  },
+  arrow_right_light: {
+    path: (
+      <path
+        d="M1 4.5H11M11 4.5L8 8M11 4.5L8 1"
+        stroke="currentColor"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    ),
+    viewBox: '0 0 12 9',
+  },
+  wyre: {
+    path: (
+      <>
+        <path
+          d="M12.0321 9.99683H18.0743L14.8068 18.0968C14.338 19.2478 13.234 20 12.0049 20H5.9624L9.22959 11.9256C9.69834 10.7746 10.8024 9.99745 12.0318 9.99745"
+          fill="#A9CAFF"
+        />
+        <path
+          d="M7.73755 15.5567H11.0932C12.3222 15.5567 12.7138 15.5077 13.7385 14.6027L18.0666 9.94019H12.0322C10.8029 9.94019 9.69567 10.7167 9.23005 11.8677L7.73755 15.5567Z"
+          fill="#0408E0"
+        />
+        <path
+          d="M10.0137 5.05469H16.0559L12.7884 13.1084C12.3196 14.2594 11.2156 14.9459 9.98652 14.9459H3.94434L7.21152 6.93656C7.68027 5.78563 8.78433 5.05469 10.0137 5.05469"
+          fill="#0384FF"
+        />
+        <path
+          d="M5.68652 10.6429H9.07902C10.3081 10.6429 10.7 10.4763 11.7243 9.57161L16.0415 5.02661H10.0137C8.78465 5.02661 7.67715 5.75567 7.21152 6.90661L5.68652 10.6429Z"
+          fill="#0408E0"
+        />
+        <path
+          d="M7.99509 0H14.0376L10.7701 8.12C10.3013 9.27094 9.19727 9.99812 7.96821 9.99812H1.92603L5.19321 1.94813C5.65884 0.7975 6.76571 0 7.99509 0Z"
+          fill="#004EFF"
+        />
+      </>
+    ),
+    viewBox: '0 0 20 20',
+  },
+  moonpay: {
+    path: (
+      <>
+        <path
+          d="M-0.000143051 18.9883C-0.000143051 27.279 6.7091 34 14.9854 34C23.2616 34 29.9709 27.279 29.9709 18.9883C29.9709 10.6975 23.2616 3.97656 14.9854 3.97656C6.7091 3.97656 -0.000143051 10.6975 -0.000143051 18.9883Z"
+          fill="black"
+        />
+        <path
+          d="M27.7887 4.07566C27.7887 6.32658 29.6103 8.15132 31.8573 8.15132C34.1043 8.15132 35.9259 6.32658 35.9259 4.07566C35.9259 1.82473 34.1043 0 31.8573 0C29.6103 0 27.7887 1.82473 27.7887 4.07566Z"
+          fill="black"
+        />
+      </>
+    ),
+    viewBox: '0 0 36 34',
+  },
+  latamex: {
+    path: (
+      <>
+        <path
+          className="st0"
+          fill="#3C2FED"
+          d="M122.1,117.9H79.2c-1.8,0-3.5-1.2-4-2.9l-14-44.6c-0.7-2.2,0.5-4.6,2.8-5.3c2.2-0.7,4.6,0.5,5.3,2.8 l13.1,41.7h39.8c2.3,0,4.2,1.9,4.2,4.2S124.4,117.9,122.1,117.9z"
+        />
+        <path
+          className="st0"
+          fill="#3C2FED"
+          d="M47,73L47,73H23.8c-2.3,0-4.2-1.9-4.2-4.2c0-2.3,1.9-4.2,4.2-4.2l0,0H47c2.3,0,4.2,1.9,4.2,4.2 C51.2,71.1,49.3,73,47,73z"
+        />
+        <path
+          className="st0"
+          fill="#3C2FED"
+          d="M53.3,95.5L53.3,95.5H23.8c-2.3,0-4.2-1.9-4.2-4.2c0-2.3,1.9-4.2,4.2-4.2l0,0h29.5c2.3,0,4.2,1.9,4.2,4.2 C57.5,93.6,55.6,95.5,53.3,95.5z"
+        />
+        <path
+          className="st0"
+          fill="#3C2FED"
+          d="M61,117.9L61,117.9H37.8c-2.3,0-4.2-1.9-4.2-4.2s1.9-4.2,4.2-4.2l0,0H61c2.3,0,4.2,1.9,4.2,4.2 S63.3,117.9,61,117.9z"
+        />
+        <path
+          className="st1"
+          fill="#3386FF"
+          d="M94.9,95.1c-1.8,0-3.5-1.1-4-2.9L77.8,50.5H37.8c-2.3,0-4.2-1.9-4.2-4.2c0-2.3,1.9-4.2,4.2-4.2h42.9 c1.8,0,3.5,1.2,4,2.9l14,44.6c0.7,2.2-0.5,4.6-2.8,5.3C95.6,95.1,95.2,95.1,94.9,95.1z"
+        />
+        <path
+          className="st1"
+          fill="#3386FF"
+          d="M136.2,95.5L136.2,95.5H113c-2.3,0-4.2-1.9-4.2-4.2c0-2.3,1.9-4.2,4.2-4.2l0,0h23.2c2.3,0,4.2,1.9,4.2,4.2 C140.4,93.6,138.4,95.5,136.2,95.5z"
+        />
+        <path
+          className="st1"
+          fill="#3386FF"
+          d="M136.2,73h-29.5c-2.3,0-4.2-1.9-4.2-4.2s1.9-4.2,4.2-4.2h29.5c2.3,0,4.2,1.9,4.2,4.2S138.4,73,136.2,73z"
+        />
+        <path
+          className="st1"
+          fill="#3386FF"
+          d="M122.1,50.5H98.9c-2.3,0-4.2-1.9-4.2-4.2c0-2.3,1.9-4.2,4.2-4.2h23.2c2.3,0,4.2,1.9,4.2,4.2 C126.4,48.6,124.4,50.5,122.1,50.5z"
+        />
+      </>
+    ),
+    viewBox: '20 10 140 140',
+  },
+  sign_transaction: {
+    path: (
+      <>
+        <path
+          d="M4.13215 13.4287L1.57348 10.8701C1.49143 10.789 1.38864 10.732 1.27636 10.7054C1.16408 10.6788 1.04666 10.6836 0.936939 10.7193C0.827221 10.755 0.729446 10.8203 0.654324 10.9078C0.579202 10.9954 0.52963 11.102 0.511039 11.2158L0.0081076 14.2712C-0.00658072 14.3617 -0.00126543 14.4544 0.0236804 14.5427C0.0486262 14.631 0.0925977 14.7128 0.152508 14.7823C0.212418 14.8518 0.286814 14.9073 0.370473 14.945C0.454132 14.9827 0.545026 15.0016 0.636772 15.0004H0.737358L3.79267 14.4975C3.90789 14.4798 4.01594 14.4305 4.10475 14.355C4.19357 14.2796 4.25964 14.1809 4.2956 14.07C4.33067 13.9585 4.33395 13.8394 4.30507 13.7261C4.2762 13.6128 4.21632 13.5099 4.13215 13.4287ZM1.0454 13.9568L1.32202 12.2783L2.72394 13.6802L1.0454 13.9568Z"
+          fill="#FDC134"
+        />
+        <path
+          d="M14.8286 2.74392L12.2735 0.182518C12.2147 0.12419 12.1449 0.078044 12.0683 0.0467254C11.9916 0.0154067 11.9095 -0.000468339 11.8267 1.05189e-05C11.6616 0.000705697 11.5033 0.0662518 11.3861 0.182518L0.687401 10.8812C0.629073 10.94 0.582927 11.0098 0.551608 11.0865C0.52029 11.1631 0.504414 11.2452 0.504893 11.3281C0.505589 11.4932 0.571135 11.6514 0.687401 11.7686L3.2488 14.33C3.36671 14.4472 3.52622 14.513 3.69248 14.513C3.85874 14.513 4.01825 14.4472 4.13616 14.33L14.8349 3.63128C14.9513 3.51254 15.0159 3.35256 15.0147 3.1863C15.0136 3.02004 14.9466 2.861 14.8286 2.74392ZM12.5944 4.09699L10.9204 2.42296L11.8267 1.523L13.4944 3.19075L12.5944 4.09699Z"
+          fill="#F08B04"
+        />
+      </>
+    ),
+    viewBox: '0 0 16 15',
+  },
+  close_squared: {
+    path: (
+      <path
+        d="M13.0024 11.328L24.328 0L26 1.67437L14.672 13.0024L26 24.328L24.328 26L13 14.672L1.67437 26L0.00236504 24.328L11.3304 13L0 1.67437L1.67437 0.0023649L13.0024 11.3304V11.328Z"
+        fill="currentColor"
+      />
+    ),
+    viewBox: '0 0 26 26',
+  },
+  send: {
+    path: (
+      <path
+        d="M6.39837 6.51154L6.88762 11.6236L11.142 12.9691C11.2651 13.0083 11.3968 13.0103 11.5209 12.9746C11.645 12.9389 11.7561 12.8671 11.8403 12.7683C11.9245 12.6694 11.9781 12.5477 11.9946 12.4182C12.0111 12.2888 11.9896 12.1573 11.9329 12.0401L6.56854 0.836288C6.51675 0.72892 6.43617 0.638428 6.33599 0.575125C6.23581 0.511823 6.12006 0.478257 6.00193 0.478257C5.88381 0.478257 5.76805 0.511823 5.66787 0.575125C5.56769 0.638428 5.48712 0.72892 5.43533 0.836288L0.0670611 12.0323C0.0103649 12.1495 -0.0110763 12.281 0.00539415 12.4104C0.0218646 12.5399 0.0755228 12.6615 0.159721 12.7604C0.243918 12.8593 0.354956 12.9311 0.479075 12.9668C0.603194 13.0024 0.73494 13.0005 0.857991 12.9612L5.11238 11.6158L5.60163 6.5135C5.61136 6.414 5.65733 6.32172 5.7306 6.25459C5.80387 6.18746 5.8992 6.15028 5.99807 6.15028C6.09693 6.15028 6.19227 6.18746 6.26553 6.25459C6.3388 6.32172 6.38477 6.414 6.3945 6.5135L6.39837 6.51154Z"
+        fill="currentColor"
+      />
+    ),
+    viewBox: '0 0 12 13',
+  },
+  copy: {
+    path: (
+      <>
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M3 5C3 3.89543 3.89543 3 5 3H12C13.1046 3 14 3.89543 14 5V14C14 15.1046 13.1046 16 12 16H5C3.89543 16 3 15.1046 3 14V5ZM12 5H5V14H12V5Z"
+          fill="currentColor"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M0 2C0 0.89543 0.895431 0 2 0H10V2L2 2V12H0V2Z"
+          fill="currentColor"
+        />
+      </>
+    ),
+    viewBox: '0 0 14 16',
+  },
+  usdc: {
+    path: (
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M256 128C256 198.692 198.692 256 128 256C57.3075 256 0 198.692 0 128C0 57.3075 57.3075 0 128 0C198.692 0 256 57.3075 256 128ZM98.8 220.8C101.6 221.7 104 220 104 217V209.4C104 207.4 102.5 205.1 100.6 204.4C69.9 193.2 48 163.8 48 129.3C48 94.9 69.9 65.4 100.6 54.1C102.5 53.4 104 51.1 104 49.1V41.6C104 38.6 101.7 36.9 98.8 37.8C60 50.2 32 86.5 32 129.3C32 172.2 60 208.4 98.8 220.8ZM132 193.3C134.2 193.3 136 191.5 136 189.3H136.1V176.5C152.6 173.9 163.2 162.4 163.2 148.2C163.2 129.5 151.8 123.2 129.8 120.2C113.5 117.9 110.4 114.1 110.4 106.5C110.4 99.3 115.9 94.2 126.5 94.2C136.1 94.2 141.6 97.6 143.9 105.3C144.4 107 145.9 108.2 147.7 108.2H156.1C158.5 108.2 160.3 106.1 159.8 103.8C157.1 91.6 148.9 84.3 136 82V69.3C136 67.1 134.2 65.3 132 65.3H124C121.8 65.3 120 67.1 120 69.3V81.7C104.2 83.9 93.9 94.5 93.9 108C93.9 125.5 104.5 132.3 126.9 135.3C142.1 137.8 146.4 141.1 146.4 149.7C146.4 158.3 139 164.1 128.7 164.1C114.7 164.1 110.1 158 108.4 150.1C108 148.3 106.4 146.9 104.5 146.9H95.4C93.1 146.9 91.3 148.9 91.7 151.2C94 164.6 102.5 174.3 120 176.7V189.3C120 191.5 121.8 193.3 124 193.3H132ZM157.2 220.8C154.3 221.7 152 220 152 217V209.5C152 207.3 153.3 205.2 155.4 204.5C186 193.3 208 163.8 208 129.4C208 94.9 186.1 65.5 155.4 54.3C153.5 53.6 152 51.3 152 49.3V41.8C152 38.8 154.4 37.1 157.2 38C196 50.2 224 86.5 224 129.3C224 172.2 196 208.4 157.2 220.8Z"
+        fill="currentColor"
+      />
+    ),
+    viewBox: '0 0 256 256',
+  },
+  twitter: {
+    path: (
+      <path
+        d="M18 2.22559C17.332 2.52441 16.6289 2.72656 15.873 2.81445C16.6465 2.35742 17.2266 1.62793 17.499 0.766602C16.7959 1.18848 15.9961 1.49609 15.1699 1.66309C14.4756 0.959961 13.5176 0.50293 12.4541 0.50293C10.415 0.50293 8.77148 2.14648 8.77148 4.18555C8.77148 4.47559 8.78906 4.77441 8.85059 5.03809C5.79199 4.87109 3.08496 3.48242 1.24805 1.25C0.94043 1.79492 0.74707 2.35742 0.74707 3.04297C0.74707 4.30859 1.39746 5.43359 2.39941 6.10156C1.79297 6.08398 1.23047 5.91699 0.729492 5.64453V5.68848C0.729492 7.48145 2.00391 8.97559 3.69141 9.30957C3.375 9.39746 3.04102 9.44141 2.70703 9.44141C2.47852 9.44141 2.25 9.41504 2.02148 9.37109C2.47852 10.8301 3.8584 11.9199 5.45801 11.9375C4.20996 12.9395 2.60156 13.6865 0.878906 13.6865C0.580078 13.6865 0.290039 13.6689 0 13.625C1.62598 14.6709 3.58594 15.1279 5.66895 15.1279C12.4541 15.1279 16.1631 9.50293 16.1631 4.625V4.1416C16.875 3.62305 17.499 2.98145 18 2.22559Z"
+        fill="#25273D"
+      />
+    ),
+    viewBox: '0 0 18 16',
+  },
+  discord: {
+    path: (
+      <>
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M7.09668 10.9676C7.09668 10.2585 7.60691 9.67725 8.25321 9.67725C8.8995 9.67725 9.42107 10.2585 9.40974 10.9676C9.40974 11.6767 8.8995 12.2579 8.25321 12.2579C7.61825 12.2579 7.09668 11.6767 7.09668 10.9676ZM11.2352 10.9677C11.2352 10.2586 11.7454 9.67742 12.3917 9.67742C13.038 9.67742 13.5483 10.2586 13.5483 10.9677C13.5483 11.6768 13.038 12.2581 12.3917 12.2581C11.7568 12.2581 11.2352 11.6768 11.2352 10.9677Z"
+          fill="#25273D"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M2.34286 0H17.6571C18.9486 0 20 1.03871 20 2.32581V22.5806L17.5429 20.4355L16.16 19.171L14.6971 17.8274L15.3029 19.9161H2.34286C1.05143 19.9161 0 18.8774 0 17.5903V2.32581C0 1.03871 1.05143 0 2.34286 0ZM12.4915 13.8307C12.8343 14.2597 13.2458 14.7452 13.2458 14.7452C15.7715 14.6662 16.7429 13.0291 16.7429 13.0291C16.7429 9.39358 15.0972 6.44681 15.0972 6.44681C13.4515 5.22745 11.8858 5.26132 11.8858 5.26132L11.7258 5.44197C13.6686 6.02907 14.5715 6.87584 14.5715 6.87584C13.3829 6.23229 12.2172 5.91616 11.1315 5.79197C10.3086 5.70165 9.52006 5.72423 8.82291 5.81455C8.76305 5.81455 8.7119 5.82315 8.65426 5.83285C8.64587 5.83426 8.63734 5.8357 8.62863 5.83713C8.22863 5.871 7.2572 6.01778 6.03434 6.54842C5.61149 6.74036 5.36006 6.87584 5.36006 6.87584C5.36006 6.87584 6.30863 5.9839 8.36577 5.39681L8.25149 5.26132C8.25149 5.26132 6.68577 5.22745 5.04006 6.44681C5.04006 6.44681 3.39434 9.39358 3.39434 13.0291C3.39434 13.0291 4.35434 14.6662 6.88006 14.7452C6.88006 14.7452 7.30291 14.2371 7.64577 13.8081C6.19434 13.3791 5.64577 12.4758 5.64577 12.4758C5.64577 12.4758 5.76006 12.5549 5.96577 12.6678C5.9772 12.6791 5.98863 12.6904 6.01149 12.7016C6.02863 12.7129 6.04577 12.7214 6.06291 12.7299C6.08006 12.7383 6.0972 12.7468 6.11434 12.7581C6.40006 12.9162 6.68577 13.0404 6.94863 13.142C7.4172 13.3226 7.9772 13.5033 8.62863 13.6275C9.48577 13.7855 10.4915 13.842 11.5886 13.6387C12.1258 13.5484 12.6743 13.3904 13.2458 13.1533C13.6458 13.0065 14.0915 12.792 14.5601 12.4871C14.5601 12.4871 13.9886 13.4129 12.4915 13.8307Z"
+          fill="#25273D"
+        />
+      </>
+    ),
+    viewBox: '0 0 20 23',
+  },
+  portis: {
+    viewBox: '0 0 300 500',
+    path: (
+      <g id="Layer_2" data-name="Layer 2">
+        <g id="Layer_1-2" data-name="Layer 1">
+          <polygon
+            fill="#133444"
+            points="4.45 262.11 15.38 257.36 150.69 198.54 296.93 262.11 241.98 375.54 107.65 394.54 6.65 285.54 4.45 262.11"
+          ></polygon>
+          <path
+            fill="#c42370"
+            d="M266.51,192.16a150.43,150.43,0,0,0-95.82-53q-1.2-.17-2.4-.3a152.36,152.36,0,0,0-35.2,0q-1.2.14-2.4.3a150.43,150.43,0,0,0-95.82,53h0l-6.08,10.08L18.66,219l-7.41,12.31v0c-.14.33-.28.67-.41,1l-.09.23,16.35,9.67,103.6,61.28,20,11.83V168.85l-20,9.08V156l20-9.09,20,9.09,101.9,46.28Z"
+          ></path>
+          <path
+            fill="#c42370"
+            d="M301.38,298.56c0,48.78-23.18,94.42-59.12,123.84a149.45,149.45,0,0,1-47.64,26.37,139.51,139.51,0,0,1-43.93,7.1C67.47,455.87,0,381.78,0,298.56a150.49,150.49,0,0,1,4.45-36.45l146.24,86.51,146.24-86.51A150.49,150.49,0,0,1,301.38,298.56Z"
+          ></path>
+          <polygon points="133.09 138.89 150.69 146.88 130.69 155.97 130.69 137.8 133.09 138.89"></polygon>
+          <polygon points="130.69 177.93 150.69 168.85 150.69 315.38 130.69 303.55 130.69 177.93"></polygon>
+          <polygon
+            fill="#1d4259"
+            points="150.69 168.85 150.69 315.38 290.82 232.49 150.69 168.85"
+          ></polygon>
+          <polygon fill="4b6b9a" points="150.69 0 150.69 168.85 290.82 232.49 150.69 0"></polygon>
+          <path
+            fill="#343535"
+            d="M150.69,168.85V315.38l-20-11.83-119.95-71-.18-.11.27-.12c.13-.33.27-.67.41-1v0l17.53-29.09L130.69,156v22Z"
+          ></path>
+          <polygon
+            fill="#3e5578"
+            points="150.69 168.85 150.69 315.38 130.69 303.55 10.74 232.6 10.56 232.49 10.83 232.37 130.69 177.93 150.69 168.85"
+          ></polygon>
+          <polygon fill="#6db2d8" points="150.69 0 150.69 168.85 10.56 232.49 150.69 0"></polygon>
+          <path
+            fill="#335f8a"
+            d="M94,443.85c2.09.92,4.21,1.79,6.34,2.61Q97.08,445.23,94,443.85Z"
+          ></path>
+          <path
+            fill="#335f8a"
+            d="M100.29,446.46c2.14.82,4.29,1.59,6.47,2.31Q103.49,447.7,100.29,446.46Z"
+          ></path>
+          <path
+            fill="#335f8a"
+            d="M112.33,450.49c1.66.47,3.33.92,5,1.33Q114.82,451.2,112.33,450.49Z"
+          ></path>
+          <path
+            fill="#6db2d8"
+            d="M4.45,262.11l146.24,86.51A167.48,167.48,0,0,1,140.22,387c-11.7,28.28-35.34,58.19-81.08,35.44h0C23.18,393,0,347.34,0,298.56A150.49,150.49,0,0,1,4.45,262.11Z"
+          ></path>
+          <path
+            fill="#529bba"
+            d="M242.24,422.41l-.36.29c-.76.62-1.52,1.23-2.29,1.83l-.2.16q-1.25,1-2.52,1.92c-.91.68-1.83,1.36-2.76,2s-1.86,1.31-2.8,1.95-1.66,1.12-2.49,1.66l-.46.3c-.88.57-1.76,1.13-2.64,1.67a2.63,2.63,0,0,1-.33.2c-.91.57-1.82,1.12-2.74,1.65s-2,1.15-3,1.7-2,1.08-3,1.6l-.11.06c-1,.52-2,1-3,1.52s-2,1-3.08,1.49-2.08,1-3.12,1.42c-2.09.92-4.21,1.79-6.34,2.61s-4.29,1.59-6.47,2.31c-.87.29-1.74.57-2.62.84s-2,.6-2.95.88c-1.66.47-3.33.92-5,1.33-.63.16-1.27.31-1.91.45l-1.07.25c-.89.2-1.78.39-2.68.57s-1.64.33-2.47.48-1.51.28-2.27.4-1.68.27-2.52.39c-.41.07-.82.13-1.24.18l-1.89.25-1.05.13c-.69.09-1.39.16-2.09.23-.86.09-1.71.17-2.57.24-.67.05-1.34.11-2,.15s-1.16.08-1.75.1h-.08c-.68,0-1.36.08-2,.1-.94,0-1.89.07-2.84.09s-1.9,0-2.86,0-1.91,0-2.86,0-1.9,0-2.84-.09c-.69,0-1.37-.06-2.05-.1h-.08c-.59,0-1.17-.06-1.75-.1s-1.34-.1-2-.15c-.86-.07-1.71-.15-2.57-.24-.7-.07-1.4-.14-2.09-.23l-1-.13c-.63-.07-1.26-.16-1.89-.25-.42,0-.83-.11-1.24-.18q-1.27-.18-2.52-.39c-.76-.12-1.52-.26-2.27-.4s-1.65-.31-2.47-.48-1.79-.37-2.68-.57l-1.07-.25c-.64-.14-1.28-.29-1.91-.45-1.68-.41-3.35-.86-5-1.33l-3-.88c-.88-.27-1.75-.55-2.62-.84-2.18-.72-4.33-1.49-6.47-2.31S96,444.77,94,443.85c-1-.46-2.08-.94-3.12-1.42s-2.06-1-3.08-1.49-2-1-3-1.52l-.11-.06c-1-.52-2-1.06-3-1.6s-2-1.12-3-1.7S76.9,435,76,434.41l-.33-.2c-.88-.54-1.76-1.1-2.64-1.67l-.46-.3c-.83-.54-1.67-1.1-2.49-1.66s-1.88-1.29-2.8-1.95-1.85-1.34-2.76-2-1.69-1.27-2.52-1.92l-.2-.16c-.77-.6-1.53-1.21-2.29-1.83l-.36-.29c45.74,22.75,69.38-7.16,81.08-35.44a167.48,167.48,0,0,0,10.47-38.35A167.47,167.47,0,0,0,161.16,387C172.86,415.25,196.5,445.16,242.24,422.41Z"
+          ></path>
+          <path
+            fill="#335f8a"
+            d="M184,451.82c1.68-.41,3.35-.86,5-1.33Q186.56,451.21,184,451.82Z"
+          ></path>
+          <path
+            fill="#335f8a"
+            d="M194.62,448.77c2.18-.72,4.33-1.49,6.47-2.31Q197.9,447.7,194.62,448.77Z"
+          ></path>
+          <path
+            fill="#335f8a"
+            d="M201.09,446.46c2.13-.82,4.25-1.69,6.34-2.61Q204.3,445.23,201.09,446.46Z"
+          ></path>
+          <path
+            fill="4b6b9a"
+            d="M301.38,298.56c0,48.78-23.19,94.43-59.14,123.85-45.74,22.75-69.38-7.16-81.08-35.44a167.47,167.47,0,0,1-10.47-38.35l146.24-86.51A150.49,150.49,0,0,1,301.38,298.56Z"
+          ></path>
+        </g>
+      </g>
+    ),
+  },
+  myetherwallet: {
+    viewBox: '0 0 100 130',
+    path: (
+      <g id="Buttons" stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
+        <g id="04.-Guide---App-icon" transform="translate(-54.000000, -188.000000)">
+          <g id="App-Icon/mew_wallet_appicon" transform="translate(54.000000, 188.000000)">
+            <path
+              d="M111.821972,8.17802805 C104.415006,0.771062258 96.991511,0.513035658 84.5656411,0.50054281 L35.4338562,0.500543063 C23.008489,0.513035658 15.5849938,0.771062258 8.17802805,8.17802805 C0.771062212,15.5849938 0.513035654,23.008489 0.50054281,35.4343589 L0.500543063,84.5661438 C0.513035654,96.991511 0.771062212,104.415006 8.17802805,111.821972 C15.9079529,119.551897 23.656924,119.5 37.0828183,119.5 L82.9171817,119.5 C96.343076,119.5 104.092047,119.551897 111.821972,111.821972 C119.551897,104.092047 119.5,96.343076 119.5,82.9171817 L119.5,37.0828183 C119.5,23.656924 119.551897,15.9079529 111.821972,8.17802805 Z"
+              id="iOS-icon-mask"
+              stroke="#DADCE0"
+              fill="#FFFFFF"
+            ></path>
+            <g id="Elements/App-Icon">
+              <g id="icon" transform="translate(2.000000, 2.000000)">
+                <path
+                  d="M78.2902062,106.530612 C100.188166,107.566326 102.396167,88.4204441 95.9815645,77.2712546 C89.7672785,66.4702342 80.4458501,62.7712544 65.3703851,64.1900465 C50.29492,65.6088385 65.4352032,90.169341 75.777958,88.3681934 C80.4702531,87.5510499 76.3984486,106.441137 78.2902062,106.530612 Z"
+                  id="right-arm"
+                  stroke="#000000"
+                  strokeWidth="2.5"
+                  fill="#FFFFFF"
+                  fillRule="nonzero"
+                ></path>
+                <path
+                  d="M37.7086826,106.530612 C15.8107234,107.566326 13.6027223,88.4204441 20.0173244,77.2712546 C26.2316101,66.4702341 35.5530388,62.7712544 50.6285038,64.1900464 C65.7039688,65.6088385 50.5636857,90.169341 40.2209306,88.3681933 C35.5286355,87.5510498 39.6004401,106.441137 37.7086826,106.530612 Z"
+                  id="left-arm"
+                  stroke="#000000"
+                  strokeWidth="2.5"
+                  fill="#FFFFFF"
+                  fillRule="nonzero"
+                ></path>
+                <path
+                  d="M60.7410433,9.46938776 C68.2263324,9.46938776 73.0871763,10.6723916 77.3062878,12.9313812 C81.5253989,15.1903709 84.8365819,18.5053464 87.0929873,22.7292899 C89.3251303,26.9078146 90.5250571,31.7150649 90.5510204,39.072678 L90.5510204,43.5436364 C90.5510204,51.0374983 89.349393,55.9039096 87.0929873,60.1278531 C84.8365819,64.3517963 81.5253989,67.6667718 77.3062878,69.9257617 C73.0871763,72.1847514 68.2263324,73.3877551 60.7410433,73.3877551 L55.2589567,73.3877551 C47.7736676,73.3877551 42.9128237,72.1847514 38.6937122,69.9257617 C34.4746011,67.6667718 31.1634181,64.3517963 28.9070125,60.1278531 C26.6506072,55.9039096 25.4489796,51.0374983 25.4489796,43.5436364 L25.4489796,39.3135064 C25.4489796,31.8196446 26.6506072,26.9532333 28.9070125,22.7292899 C31.1634181,18.5053464 34.4746011,15.1903709 38.6937122,12.9313812 C42.8674567,10.6966818 47.6692075,9.4953809 55.0184035,9.46938776 L60.7410433,9.46938776 Z"
+                  id="helmet"
+                  stroke="#000000"
+                  strokeWidth="2.5"
+                  fill="#FFFFFF"
+                  fillRule="nonzero"
+                ></path>
+                <path
+                  d="M54.0368739,27.8163265 L61.9631261,27.8163265 C65.9787611,27.8163265 69.0898725,28.5567899 71.7010112,29.9472255 C74.3121499,31.3376607 76.361387,33.3780684 77.7578394,35.9779566 C79.1542915,38.5778448 79.8979592,41.6755514 79.8979592,45.6738844 L79.8979592,46.6526462 C79.8979592,50.6509792 79.1542915,53.7486858 77.7578394,56.348574 C76.361387,58.9484622 74.3121499,60.9888699 71.7010112,62.3793051 C69.0898725,63.7697406 65.9787611,64.5102041 61.9631261,64.5102041 L54.0368739,64.5102041 C50.0212389,64.5102041 46.9101275,63.7697406 44.2989888,62.3793051 C41.6878501,60.9888699 39.638613,58.9484622 38.2421606,56.348574 C36.8457085,53.7486858 36.1020408,50.6509792 36.1020408,46.6526462 L36.1020408,45.6738844 C36.1020408,41.6755514 36.8457085,38.5778448 38.2421606,35.9779566 C39.638613,33.3780684 41.6878501,31.3376607 44.2989888,29.9472255 C46.9101275,28.5567899 50.0212389,27.8163265 54.0368739,27.8163265 Z"
+                  id="screen"
+                  stroke="#000000"
+                  strokeWidth="2.5"
+                  fill="#FFFFFF"
+                  fillRule="nonzero"
+                ></path>
+                <circle
+                  id="left-eye"
+                  fill="#000000"
+                  fillRule="nonzero"
+                  cx="44.0918367"
+                  cy="46.4591837"
+                  r="4.73469388"
+                ></circle>
+                <circle
+                  id="right-eye"
+                  fill="#000000"
+                  fillRule="nonzero"
+                  cx="71.9081633"
+                  cy="46.4591837"
+                  r="4.73469388"
+                ></circle>
+                <path
+                  d="M52.9693878,45.5714286 C52.9693878,48.1863278 55.2216695,50.3061224 58,50.3061224 C60.7783305,50.3061224 63.0306122,48.1863278 63.0306122,45.5714286"
+                  id="month"
+                  stroke="#000000"
+                  strokeWidth="2.5"
+                  strokeLinecap="round"
+                ></path>
+                <g id="left-ear" transform="translate(16.867347, 27.816327)">
+                  <path
+                    d="M8.50342442,0.0863375259 C9.91857046,0.286175668 11.1461997,0.798351267 12.1319865,1.5504265 C13.1177733,2.30250175 13.8617176,3.29447664 14.309494,4.45391321 C14.7572701,5.61334978 14.9088781,6.94024834 14.7099923,8.36217071 L14.5288878,9.65696686 C14.3300019,11.0788892 13.8202671,12.3123968 13.071776,13.302904 C12.3232849,14.2934111 11.3360374,15.0409178 10.1821262,15.4908383 C9.02821495,15.9407585 7.70764003,16.0930925 6.29249399,15.8932544 C4.87734794,15.6934163 3.64971867,15.1812405 2.66393187,14.4291654 C1.67814507,13.6770901 0.93420071,12.6851152 0.486424508,11.5256786 C0.0386483057,10.366242 -0.112959744,9.0393435 0.0859260681,7.61742112 L0.267030596,6.32262498 C0.465916407,4.9007026 0.975651203,3.667195 1.72414235,2.67668787 C2.4726335,1.6861807 3.459881,0.93867404 4.6137922,0.488753688 C5.76770343,0.0388333657 7.08827838,-0.113500645 8.50342442,0.0863375259 Z"
+                    id="ear"
+                    stroke="#000000"
+                    strokeWidth="2.5"
+                    fill="#FFFFFF"
+                    fillRule="nonzero"
+                  ></path>
+                  <line
+                    x1="6.37952543"
+                    y1="5.0274423"
+                    x2="6.31931734"
+                    y2="9.34663542"
+                    id="Path-3"
+                    stroke="#D8D8D8"
+                    strokeWidth="3.25510204"
+                    strokeLinecap="round"
+                    strokeLinejoin="round"
+                    transform="translate(6.349421, 7.187039) rotate(8.000000) translate(-6.349421, -7.187039) "
+                  ></line>
+                </g>
+                <g id="right-ear" transform="translate(84.336735, 27.816327)">
+                  <path
+                    d="M6.2618911,0.0841025481 C4.85094235,0.278035669 3.62831666,0.773308529 2.64727023,1.50080271 C1.6662238,2.22829697 0.926756924,3.18801232 0.482125803,4.31083077 C0.0374943827,5.43364922 -0.112301284,6.71957077 0.0859944004,8.09947658 L0.33833548,9.85547156 C0.536631465,11.2353777 1.04304614,12.4311036 1.78690638,13.3905656 C2.53076662,14.3500274 3.51207212,15.073225 4.66015005,15.5080737 C5.80822797,15.9429225 7.12307803,16.0894225 8.53402708,15.8954892 C9.94497612,15.7015561 11.1676018,15.2062832 12.1486482,14.4787892 C13.1296944,13.7512949 13.8691612,12.7915795 14.3137927,11.6687611 C14.7584241,10.5459426 14.9082198,9.26002107 14.7099238,7.88011526 L14.457583,6.12412028 C14.259287,4.74421417 13.7528723,3.54848824 13.0090121,2.5890262 C12.2651519,1.62956451 11.2838461,0.906367009 10.1357681,0.471518224 C8.9876902,0.0366694379 7.67284015,-0.109830602 6.2618911,0.0841025481 Z"
+                    id="Rectangle"
+                    stroke="#000000"
+                    strokeWidth="2.5"
+                    fill="#FFFFFF"
+                    fillRule="nonzero"
+                  ></path>
+                  <line
+                    x1="8.46951223"
+                    y1="5.02744227"
+                    x2="8.40930409"
+                    y2="9.34663539"
+                    id="Path-3"
+                    stroke="#D8D8D8"
+                    strokeWidth="3.25510204"
+                    strokeLinecap="round"
+                    strokeLinejoin="round"
+                    transform="translate(8.439408, 7.187039) scale(-1, 1) rotate(8.000000) translate(-8.439408, -7.187039) "
+                  ></line>
+                </g>
+                <g id="logo" transform="translate(32.255102, 58.591837)" fillRule="nonzero">
+                  <path
+                    d="M25.744898,0 C11.5257474,0 0,11.526868 0,25.744898 C0,39.962928 11.5268702,51.4897959 25.744898,51.4897959 C39.9629259,51.4897959 51.4897959,39.962928 51.4897959,25.744898 C51.4897959,11.526868 39.9646053,0 25.744898,0 Z"
+                    id="Path"
+                    fill="#07A6B3"
+                  ></path>
+                  <path
+                    d="M25.744898,0 C11.5257474,0 0,11.526868 0,25.744898 C0,39.962928 11.5268702,51.4897959 25.744898,51.4897959 C39.9629259,51.4897959 51.4897959,39.962928 51.4897959,25.744898 C51.4897959,11.526868 39.9646053,0 25.744898,0 Z"
+                    id="Path"
+                    stroke="#000000"
+                    strokeWidth="2.5"
+                  ></path>
+                  <path
+                    d="M10.5747318,19.933241 L17.5712648,22.9090435 C17.54866,23.0066523 17.5198385,23.1347351 17.4876233,23.278616 C17.2782827,23.9600935 17.155078,24.6650236 17.1208711,25.3770231 C17.1057594,25.5181127 17.1004708,25.6600752 17.1050536,25.8018938 L17.1112704,25.8995026 L17.1112704,25.8995026 C17.1426831,28.5013969 18.3546872,30.9485523 20.4064807,32.5528721 C22.4582839,34.1571827 25.1282665,34.7453775 27.6656853,34.1520604 L31.6428772,40.8952791 C29.7702781,41.6410616 27.77227,42.0229311 25.7562013,42.0203956 C16.987505,42.0203956 9.46938776,34.9228115 9.46938776,26.1652602 C9.46938776,24.1091733 9.8609994,21.7359824 10.5747318,19.933241 Z M25.7578933,9.46940335 C34.5265898,9.46940335 42.0204082,16.6984625 42.0204082,25.4537595 C42.0188219,27.6680417 41.7505206,29.8773667 40.9528869,31.8041756 L40.7853878,32.1843878 L23.0256268,24.6711529 L34.4175299,24.6215044 C34.0083913,20.4957866 29.7616442,17.1419515 25.7567718,17.1419515 C24.8485591,17.1411745 23.9465535,17.2914341 23.087785,17.5865722 L19.1462021,10.9048581 C21.2206294,9.95587716 23.4761075,9.46620246 25.7578933,9.46940335 Z"
+                    id="Combined-Shape"
+                    fill="#FFFFFF"
+                  ></path>
+                </g>
+                <g
+                  id="hand"
+                  transform="translate(37.383915, 95.943703) rotate(-1.000000) translate(-37.383915, -95.943703) translate(27.883915, 84.943703)"
+                  fill="#FFFFFF"
+                  fillRule="nonzero"
+                  stroke="#000000"
+                  strokeWidth="2.5"
+                >
+                  <path
+                    d="M2.49642257,16.0564812 L0.670687359,10.6382004 C-0.719178706,6.51345957 1.54691405,2.05927254 5.73214719,0.689495241 L5.82626625,0.659328139 L5.82626625,0.659328139 C10.0961729,-0.680484311 14.6654895,1.61250088 16.0777701,5.80376189 L17.9035053,11.2220427 C19.2933714,15.3467835 17.0272786,19.8009706 12.8420455,21.1707479 L12.7479264,21.200915 L12.7479264,21.200915 C8.47801977,22.5407274 3.90870315,20.2477422 2.49642257,16.0564812 Z"
+                    id="RectangleCopy-3"
+                  ></path>
+                  <line
+                    x1="12.1092443"
+                    y1="9.54323997"
+                    x2="12.1513308"
+                    y2="3.32969571"
+                    id="LineCopy-2"
+                    strokeLinecap="round"
+                    transform="translate(12.130288, 6.436468) rotate(-108.000000) translate(-12.130288, -6.436468) "
+                  ></line>
+                  <line
+                    x1="14.307877"
+                    y1="14.6120534"
+                    x2="14.2684898"
+                    y2="8.39789246"
+                    id="LineCopy-3"
+                    strokeLinecap="round"
+                    transform="translate(14.288183, 11.504973) rotate(-108.000000) translate(-14.288183, -11.504973) "
+                  ></line>
+                </g>
+                <g
+                  id="hand"
+                  transform="translate(78.680251, 95.943703) scale(-1, 1) rotate(-1.000000) translate(-78.680251, -95.943703) translate(69.180251, 84.943703)"
+                  fill="#FFFFFF"
+                  fillRule="nonzero"
+                  stroke="#000000"
+                  strokeWidth="2.5"
+                >
+                  <path
+                    d="M2.49642257,16.0564812 L0.670687359,10.6382004 C-0.719178706,6.51345957 1.54691405,2.05927254 5.73214719,0.689495241 L5.82626625,0.659328139 L5.82626625,0.659328139 C10.0961729,-0.680484311 14.6654895,1.61250088 16.0777701,5.80376189 L17.9035053,11.2220427 C19.2933714,15.3467835 17.0272786,19.8009706 12.8420455,21.1707479 L12.7479264,21.200915 L12.7479264,21.200915 C8.47801977,22.5407274 3.90870315,20.2477422 2.49642257,16.0564812 Z"
+                    id="RectangleCopy-3"
+                  ></path>
+                  <line
+                    x1="12.1092443"
+                    y1="9.54323997"
+                    x2="12.1513308"
+                    y2="3.32969571"
+                    id="LineCopy-2"
+                    strokeLinecap="round"
+                    transform="translate(12.130288, 6.436468) rotate(-108.000000) translate(-12.130288, -6.436468) "
+                  ></line>
+                  <line
+                    x1="14.307877"
+                    y1="14.6120534"
+                    x2="14.2684898"
+                    y2="8.39789246"
+                    id="LineCopy-3"
+                    strokeLinecap="round"
+                    transform="translate(14.288183, 11.504973) rotate(-108.000000) translate(-14.288183, -11.504973) "
+                  ></line>
+                </g>
+              </g>
+            </g>
+          </g>
+        </g>
+      </g>
+    ),
+  },
+  gnosis_safe: {
+    viewBox: '0 0 270 280',
+    path: (
+      <path d="M137,9.84A128.75,128.75,0,1,0,265.7,138.59,128.76,128.76,0,0,0,137,9.84Zm94.23,135.78H171.44a36.38,36.38,0,1,1,.28-12.66h59.46a6.33,6.33,0,0,1,0,12.66Z" />
+    ),
+  },
+  plus: {
+    path: (
+      <>
+        <rect x="9" y="6" width="2" height="8" rx="1" fill="currentColor" />
+        <rect
+          x="14"
+          y="9"
+          width="2"
+          height="8"
+          rx="1"
+          transform="rotate(90 14 9)"
+          fill="currentColor"
+        />
+      </>
+    ),
+    viewBox: '0 0 20 20',
+  },
+  plus_header: {
+    path: (
+      <path
+        d="M10 1C10 0.447715 9.55229 5.96046e-08 9 0C8.44771 0 8 0.447716 8 1V8H1C0.447715 8 0 8.44771 0 9C0 9.55229 0.447716 10 1 10H8V17C8 17.5523 8.44771 18 9 18C9.55228 18 10 17.5523 10 17V10H17C17.5523 10 18 9.55229 18 9C18 8.44771 17.5523 8 17 8H10V1Z"
+        fill="currentColor"
+      />
+    ),
+    viewBox: '0 0 18 18',
+  },
+  minus: {
+    path: (
+      <rect
+        x="14"
+        y="9"
+        width="2"
+        height="8"
+        rx="1"
+        transform="rotate(90 14 9)"
+        fill="currentColor"
+      />
+    ),
+    viewBox: '0 0 20 20',
+  },
+  tooltip: {
+    path: (
+      <>
+        <circle cx="7" cy="7" r="6.3" stroke="currentColor" strokeWidth="1.4" fill="none" />
+        <path
+          d="M6.12953 11H7.86603V5.54545H6.12953V11ZM6.99956 4.9098C7.49316 4.9098 7.89089 4.53693 7.89089 4.07884C7.89089 3.62074 7.49316 3.25142 6.99956 3.25142C6.5095 3.25142 6.10822 3.62074 6.10822 4.07884C6.10822 4.53693 6.5095 4.9098 6.99956 4.9098Z"
+          fill="currentColor"
+        />
+      </>
+    ),
+    viewBox: '0 0 14 14',
+  },
+}
+
+const customLandingIcons = {
+  bannerWallet: {
+    path: (
+      <svg
+        width="24"
+        height="24"
+        viewBox="0 0 24 24"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <path
+          d="M2 6C2 3.79086 3.79086 2 6 2H18C19.1046 2 20 2.89543 20 4V7H2V6Z"
+          stroke="url(#paint0_linear)"
+          strokeWidth="2.5"
+        />
+        <path
+          d="M2 7H19C20.6569 7 22 8.34315 22 10V19C22 20.6569 20.6569 22 19 22H6C3.79086 22 2 20.2091 2 18V7Z"
+          stroke="url(#paint1_linear)"
+          strokeWidth="2.5"
+        />
+        <circle cx="17.5" cy="14.5" r="1.5" fill="url(#paint2_linear)" />
+        <defs>
+          <linearGradient
+            id="paint0_linear"
+            x1="11"
+            y1="2"
+            x2="11"
+            y2="7"
+            gradientUnits="userSpaceOnUse"
+          >
+            <stop stopColor="#282A42" />
+            <stop offset="1" stopColor="#4F548B" />
+          </linearGradient>
+          <linearGradient
+            id="paint1_linear"
+            x1="12"
+            y1="7"
+            x2="12"
+            y2="22"
+            gradientUnits="userSpaceOnUse"
+          >
+            <stop stopColor="#282A42" />
+            <stop offset="1" stopColor="#4F548B" />
+          </linearGradient>
+          <linearGradient
+            id="paint2_linear"
+            x1="15.9483"
+            y1="15.7658"
+            x2="18.5873"
+            y2="13.3808"
+            gradientUnits="userSpaceOnUse"
+          >
+            <stop stopColor="#25273D" />
+            <stop offset="1" stopColor="#4F558C" />
+          </linearGradient>
+        </defs>
+      </svg>
+    ),
+    viewBox: '0 0 24 24',
+  },
+  exclamationMark: {
+    path: (
+      <svg width="6" height="24" viewBox="0 0 6 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M3 0C1.41149 0 0.186548 1.39911 0.396491 2.97368L1.86784 14.0088C1.9435 14.5762 2.42753 15 3 15C3.57247 15 4.0565 14.5762 4.13216 14.0088L5.60351 2.97368C5.81345 1.39911 4.58851 0 3 0ZM3 24C4.65685 24 6 22.6569 6 21C6 19.3431 4.65685 18 3 18C1.34315 18 0 19.3431 0 21C0 22.6569 1.34315 24 3 24Z"
+          fill="currentColor"
+        />
+      </svg>
+    ),
+    viewBox: '0 0 6 24',
+  },
+  announcement: {
+    path: (
+      <path
+        d="M21.5725 11.1185C21.5725 10.4553 21.5655 10.4132 21.5717 9.42297V5.78125C21.5717 4.65943 20.667 3.75 19.551 3.75C17.4991 5.52735 15.0665 7.81251 10.9629 7.81251H5.91105C5.07403 7.81251 4.39551 8.49456 4.39551 9.33595V12.3828C4.39551 13.2242 5.07403 13.9063 5.91105 13.9063H7.27704C6.68528 15.9581 7.35531 17.5999 7.87107 19.2205C7.92481 19.3894 8.03535 19.5345 8.18434 19.6293C8.84045 20.0468 10.3426 20.1262 11.1451 19.7928C11.6518 19.5823 11.7753 18.9224 11.3826 18.538C10.7983 17.9659 10.4311 17.3665 10.7576 16.6064C10.7833 16.5468 10.7937 16.4817 10.7879 16.417C10.7822 16.3523 10.7604 16.29 10.7246 16.236C10.2097 15.4527 10.6116 14.2648 11.625 13.9269C15.3328 14.1566 17.6121 16.2892 19.551 17.9688C20.667 17.9688 21.5717 17.0593 21.5717 15.9375V12.2958C21.5725 11.1185 21.5717 12.2958 21.5725 11.5346V11.1185ZM19.551 15.2982C17.8268 13.8706 15.6032 12.3484 12.4784 11.9662V9.75254C15.6032 9.37035 17.8268 7.84818 19.551 6.42059V15.2982Z"
+        fill="#25273D"
+      />
+    ),
+    viewBox: '0 0 25 25',
+  },
+  landing_card_dai: {
+    path: (
+      <>
+        <circle cx="44" cy="44" r="44" fill="url(#paint0_linear_landing_card_dai)" />
+        <circle cx="43.9999" cy="43.9997" r="36.6667" fill="url(#paint1_linear_landing_card_dai)" />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M30.0938 26.7324H44.0613C52.5572 26.7324 58.9973 31.299 61.3932 37.9438H65.7441V41.9601H62.3094C62.3767 42.5947 62.4109 43.2413 62.4109 43.8983V43.9969C62.4109 44.7365 62.3675 45.4636 62.2826 46.1756H65.7441V50.1919H61.3103C58.8514 56.7443 52.4612 61.2673 44.0613 61.2673H30.0938V50.1919H25.2402V46.1756H30.0938V41.9601H25.2402V37.9438H30.0938V26.7324ZM33.9985 50.1919V57.6645H44.0613C50.2709 57.6645 54.8843 54.6716 57.0319 50.1919H33.9985ZM58.2282 46.1756H33.9985V41.9601H58.2344C58.3242 42.6235 58.3701 43.3035 58.3701 43.9969V44.0956C58.3701 44.8049 58.3221 45.4994 58.2282 46.1756ZM44.0613 30.3294C50.2995 30.3294 54.9268 33.4012 57.0614 37.9438H33.9985V30.3294H44.0613Z"
+          fill="white"
+        />
+        <defs>
+          <linearGradient
+            id="paint0_linear_landing_card_dai"
+            x1="44"
+            y1="0"
+            x2="44"
+            y2="88"
+            gradientUnits="userSpaceOnUse"
+          >
+            <stop stopColor="#FEDB88" />
+            <stop offset="1" stopColor="#FDC134" />
+          </linearGradient>
+          <linearGradient
+            id="paint1_linear_landing_card_dai"
+            x1="43.9999"
+            y1="-3.05588"
+            x2="43.9999"
+            y2="88.3663"
+            gradientUnits="userSpaceOnUse"
+          >
+            <stop stopColor="#F9A606" />
+            <stop offset="1" stopColor="#FBCC5F" />
+          </linearGradient>
+        </defs>
+      </>
+    ),
+    viewBox: '0 0 88 88',
+  },
+  landing_card_faq: {
+    path: (
+      <>
+        <circle cx="44" cy="44" r="44" fill="url(#paint0_card_faq)" />
+        <g style={{ mixBlendMode: 'overlay' }} opacity="0.5">
+          <circle cx="43.9999" cy="44.0002" r="36.6667" fill="black" />
+        </g>
+        <path
+          d="M40.6667 51.2592H45.1112V51.037C45.1853 46.4444 46.3704 44.4444 49.6297 42.4074C52.889 40.4259 54.8149 37.574 54.8149 33.4814C54.8149 27.7037 50.5927 23.4814 44.0741 23.4814C38.0741 23.4814 33.3149 27.1852 33.0371 33.4814H37.7038C37.9816 29.1111 41.0371 27.3333 44.0741 27.3333C47.5556 27.3333 50.3704 29.6296 50.3704 33.2592C50.3704 36.2037 48.6853 38.3148 46.5186 39.6296C42.889 41.8333 40.7223 43.9814 40.6667 51.037V51.2592ZM43.0371 62.2222C44.8704 62.2222 46.3704 60.7222 46.3704 58.8889C46.3704 57.0555 44.8704 55.5555 43.0371 55.5555C41.2038 55.5555 39.7038 57.0555 39.7038 58.8889C39.7038 60.7222 41.2038 62.2222 43.0371 62.2222Z"
+          fill="white"
+        />
+        <defs>
+          <radialGradient
+            id="paint0_card_faq"
+            cx="0"
+            cy="0"
+            r="1"
+            gradientUnits="userSpaceOnUse"
+            gradientTransform="translate(9.77778 19.5556) rotate(40.8335) scale(87.2309 268.794)"
+          >
+            <stop stopColor="#FFCCEB" />
+            <stop offset="0.473958" stopColor="#BFC6FF" />
+            <stop offset="1" stopColor="#96E4D2" />
+          </radialGradient>
+        </defs>
+      </>
+    ),
+    viewBox: '0 0 88 88',
+  },
+  support_plus: {
+    path: (
+      <path
+        d="M9.42857 19.7143C9.42857 20.1877 9.81233 20.5714 10.2857 20.5714C10.7591 20.5714 11.1429 20.1877 11.1429 19.7143V11.1429H19.7143C20.1877 11.1429 20.5714 10.7591 20.5714 10.2857C20.5714 9.81233 20.1877 9.42857 19.7143 9.42857H11.1429V0.857143C11.1429 0.383756 10.7591 0 10.2857 0C9.81232 0 9.42857 0.383756 9.42857 0.857143V9.42857H0.857143C0.383756 9.42857 0 9.81233 0 10.2857C0 10.7591 0.383756 11.1429 0.857143 11.1429H9.42857V19.7143Z"
+        fill="currentColor"
+      />
+    ),
+    viewBox: '0 0 21 21',
+  },
+  support_minus: {
+    path: (
+      <path
+        d="M9.42857 2.64295C9.42857 3.11634 9.81232 2.64295 10.2857 2.64295C10.7591 2.64295 11.1429 3.11634 11.1429 2.64295H19.7143C20.1877 2.64295 20.5714 2.2592 20.5714 1.78581C20.5714 1.31242 20.1877 0.928667 19.7143 0.928667H11.1429L14.5 1.49993C14.5 1.02654 10.7591 0.928406 10.2857 0.928406C10.2857 0.928406 6 1.02655 6 1.49994L9.42857 0.928667H0.857143C0.383756 0.928667 0 1.31242 0 1.78581C0 2.2592 0.383756 2.64295 0.857143 2.64295H9.42857Z"
+        fill="currentColor"
+      />
+    ),
+    viewBox: '0 0 21 3',
+  },
+  multiply_transition: {
+    path: (
+      <>
+        <rect x="78" y="28" width="230" height="163" rx="16" fill="url(#paint0_linear)" />
+        <rect x="78" y="28" width="230" height="163" rx="16" fill="url(#paint1_radial)" />
+        <rect x="78" y="28" width="230" height="163" rx="16" fill="url(#paint2_radial)" />
+        <g filter="url(#filter0_d)">
+          <rect x="8" y="43" width="140" height="134" rx="8" fill="white" />
+        </g>
+        <rect x="22" y="57" width="113" height="22" rx="11" fill="#F3F7F9" />
+        <g filter="url(#filter1_d)">
+          <rect x="22" y="57" width="44" height="22" rx="11" fill="white" />
+        </g>
+        <rect x="22.5" y="94.5" width="112" height="25" rx="3.5" stroke="#EAEAEA" />
+        <circle cx="34" cy="107" r="2" fill="#787A9B" />
+        <circle cx="43" cy="107" r="2" fill="#787A9B" />
+        <circle cx="52" cy="107" r="2" fill="#787A9B" />
+        <rect opacity="0.8" x="22" y="134" width="110" height="20" rx="10" fill="#25273D" />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M209.814 113.084C210.388 112.686 210.388 111.837 209.814 111.44L197.569 102.957C196.906 102.497 196 102.972 196 103.779V120.744C196 121.551 196.906 122.026 197.569 121.566L209.814 113.084ZM187 115.26C188.657 115.26 190 113.917 190 112.26C190 110.604 188.657 109.26 187 109.26C185.343 109.26 184 110.604 184 112.26C184 113.917 185.343 115.26 187 115.26ZM179 112.26C179 113.917 177.657 115.26 176 115.26C174.343 115.26 173 113.917 173 112.26C173 110.604 174.343 109.26 176 109.26C177.657 109.26 179 110.604 179 112.26ZM165 115.26C166.657 115.26 168 113.917 168 112.26C168 110.604 166.657 109.26 165 109.26C163.343 109.26 162 110.604 162 112.26C162 113.917 163.343 115.26 165 115.26Z"
+          fill="url(#paint3_linear)"
+        />
+        <g filter="url(#filter2_d)">
+          <rect x="222" y="8" width="190" height="203" rx="8" fill="white" />
+        </g>
+        <rect opacity="0.8" x="240" y="171" width="154" height="22" rx="11" fill="#25273D" />
+        <g
+          style={{
+            mixBlendMode: 'hard-light',
+          }}
+        >
+          <path
+            d="M222 16C222 11.5817 225.582 8 230 8H404C408.418 8 412 11.5817 412 16V19H222V16Z"
+            fill="url(#paint4_linear)"
+          />
+        </g>
+        <rect x="240" y="36" width="154" height="26" rx="13" fill="#F3F7F9" />
+        <g filter="url(#filter3_d)">
+          <rect x="240" y="36" width="77" height="26" rx="13" fill="white" />
+        </g>
+        <rect x="240" y="134" width="27" height="3" rx="1.5" fill="#EAEAEA" />
+        <rect x="375" y="134" width="19" height="3" rx="1.5" fill="#BEC9D0" />
+        <rect x="240" y="143" width="38" height="3" rx="1.5" fill="#EAEAEA" />
+        <rect x="387" y="143" width="7" height="3" rx="1.5" fill="#BEC9D0" />
+        <rect x="240" y="152" width="27" height="3" rx="1.5" fill="#EAEAEA" />
+        <rect x="381" y="152" width="13" height="3" rx="1.5" fill="#BEC9D0" />
+        <rect x="240" y="78" width="18" height="3" rx="1.5" fill="#EAEAEA" />
+        <rect x="376" y="78" width="18" height="3" rx="1.5" fill="#EAEAEA" />
+        <rect x="240" y="85" width="27" height="3" rx="1.5" fill="#787A9B" />
+        <rect x="367" y="85" width="27" height="3" rx="1.5" fill="#787A9B" />
+        <rect x="239" y="109" width="155" height="3" rx="1.5" fill="#EAEAEA" />
+        <rect x="239" y="109" width="81" height="3" rx="1.5" fill="#D8D9FE" />
+        <circle cx="317" cy="111" r="5" fill="url(#paint5_linear)" />
+        <defs>
+          <filter
+            id="filter0_d"
+            x="0"
+            y="35"
+            width="156"
+            height="150"
+            filterUnits="userSpaceOnUse"
+            colorInterpolationFilters="sRGB"
+          >
+            <feFlood floodOpacity="0" result="BackgroundImageFix" />
+            <feColorMatrix
+              in="SourceAlpha"
+              type="matrix"
+              values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
+              result="hardAlpha"
+            />
+            <feOffset />
+            <feGaussianBlur stdDeviation="4" />
+            <feColorMatrix type="matrix" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.15 0" />
+            <feBlend mode="normal" in2="BackgroundImageFix" result="effect1_dropShadow" />
+            <feBlend mode="normal" in="SourceGraphic" in2="effect1_dropShadow" result="shape" />
+          </filter>
+          <filter
+            id="filter1_d"
+            x="18"
+            y="53"
+            width="52"
+            height="30"
+            filterUnits="userSpaceOnUse"
+            colorInterpolationFilters="sRGB"
+          >
+            <feFlood floodOpacity="0" result="BackgroundImageFix" />
+            <feColorMatrix
+              in="SourceAlpha"
+              type="matrix"
+              values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
+              result="hardAlpha"
+            />
+            <feOffset />
+            <feGaussianBlur stdDeviation="2" />
+            <feColorMatrix type="matrix" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.15 0" />
+            <feBlend mode="normal" in2="BackgroundImageFix" result="effect1_dropShadow" />
+            <feBlend mode="normal" in="SourceGraphic" in2="effect1_dropShadow" result="shape" />
+          </filter>
+          <filter
+            id="filter2_d"
+            x="214"
+            y="0"
+            width="206"
+            height="219"
+            filterUnits="userSpaceOnUse"
+            colorInterpolationFilters="sRGB"
+          >
+            <feFlood floodOpacity="0" result="BackgroundImageFix" />
+            <feColorMatrix
+              in="SourceAlpha"
+              type="matrix"
+              values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
+              result="hardAlpha"
+            />
+            <feOffset />
+            <feGaussianBlur stdDeviation="4" />
+            <feColorMatrix type="matrix" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.15 0" />
+            <feBlend mode="normal" in2="BackgroundImageFix" result="effect1_dropShadow" />
+            <feBlend mode="normal" in="SourceGraphic" in2="effect1_dropShadow" result="shape" />
+          </filter>
+          <filter
+            id="filter3_d"
+            x="236"
+            y="32"
+            width="85"
+            height="34"
+            filterUnits="userSpaceOnUse"
+            colorInterpolationFilters="sRGB"
+          >
+            <feFlood floodOpacity="0" result="BackgroundImageFix" />
+            <feColorMatrix
+              in="SourceAlpha"
+              type="matrix"
+              values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
+              result="hardAlpha"
+            />
+            <feOffset />
+            <feGaussianBlur stdDeviation="2" />
+            <feColorMatrix type="matrix" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.15 0" />
+            <feBlend mode="normal" in2="BackgroundImageFix" result="effect1_dropShadow" />
+            <feBlend mode="normal" in="SourceGraphic" in2="effect1_dropShadow" result="shape" />
+          </filter>
+          <linearGradient
+            id="paint0_linear"
+            x1="181.409"
+            y1="191"
+            x2="182.322"
+            y2="28.0026"
+            gradientUnits="userSpaceOnUse"
+          >
+            <stop stopColor="#F2EDFF" />
+            <stop offset="0.40625" stopColor="#F6EEF9" />
+            <stop offset="1" stopColor="#FCEFF1" />
+          </linearGradient>
+          <radialGradient
+            id="paint1_radial"
+            cx="0"
+            cy="0"
+            r="1"
+            gradientUnits="userSpaceOnUse"
+            gradientTransform="translate(80.3469 28) rotate(29.1405) scale(260.643 615.79)"
+          >
+            <stop stopColor="#BFFFF0" />
+            <stop offset="0.336378" stopColor="#FEFFC7" stopOpacity="0" />
+          </radialGradient>
+          <radialGradient
+            id="paint2_radial"
+            cx="0"
+            cy="0"
+            r="1"
+            gradientUnits="userSpaceOnUse"
+            gradientTransform="translate(80.3469 28) rotate(29.1405) scale(260.643 615.79)"
+          >
+            <stop offset="0.666667" stopColor="#FEFFC7" stopOpacity="0" />
+            <stop offset="1" stopColor="#CEFFE6" stopOpacity="0.767746" />
+          </radialGradient>
+          <linearGradient
+            id="paint3_linear"
+            x1="162"
+            y1="122"
+            x2="206.568"
+            y2="121.983"
+            gradientUnits="userSpaceOnUse"
+          >
+            <stop stopColor="#25273D" stopOpacity="0.3" />
+            <stop offset="1" stopColor="#25273D" />
+          </linearGradient>
+          <linearGradient
+            id="paint4_linear"
+            x1="316.159"
+            y1="-0.555554"
+            x2="345.867"
+            y2="128.439"
+            gradientUnits="userSpaceOnUse"
+          >
+            <stop stopColor="#5E5AFF" />
+            <stop offset="0.0001" stopColor="#5E5AFF" />
+            <stop offset="1" stopColor="#30C1FF" />
+          </linearGradient>
+          <linearGradient
+            id="paint5_linear"
+            x1="312"
+            y1="106"
+            x2="321.55"
+            y2="116.25"
+            gradientUnits="userSpaceOnUse"
+          >
+            <stop stopColor="#2A30EE" />
+            <stop offset="1" stopColor="#A4A6FF" />
+          </linearGradient>
+        </defs>
+      </>
+    ),
+    viewBox: '0 0 420 219',
+  },
+}
+
+export const icons = {
+  ...daiUiIcons,
+  ...brandingIcons,
+  ...customIcons,
+  ...customLandingIcons,
+  ...tokenIcons,
+}
